+++ .docker/start.sh
@@ -3,7 +3,7 @@ export $(grep -v '^#' .env | xargs)
 
 FILE="docker-compose-qgis.yml"
 
-docker-compose -f ${FILE} up -d --force-recreate --remove-orphans
+docker compose -f ${FILE} up -d --force-recreate --remove-orphans
 echo "Wait 10 seconds"
 sleep 10
 echo "Installation of the plugin ${PLUGIN_NAME}"